refactor(py): moved ontologies to futures and adatpers to nav_msgs.py
