+++ mosaico-sdk-py/src/mosaicolabs/models/futures/__init__.py
@@ -3,6 +3,11 @@ from .depth_camera import (
     StereoCamera as StereoCamera,
     ToFCamera as ToFCamera,
 )
+from .grids import (
+    GridCells as GridCells,
+    MapMetadata as MapMetadata,
+    OccupancyGrid as OccupancyGrid,
+)
 from .laser import (
     LaserScan as LaserScan,
     MultiEchoLaserScan as MultiEchoLaserScan,
